renamed TestCase.java to ApprovalTestCase for clarity of intent and moved TestUDPClient to new testUtilities package
