+++ jpo-ode-svcs/src/test/java/us/dot/its/jpo/ode/testUtilities/ApprovalTestCase.java
@@ -1,4 +1,4 @@
-package us.dot.its.jpo.ode.udp.map;
+package us.dot.its.jpo.ode.testUtilities;
 
 import lombok.Getter;
 import lombok.Setter;
@@ -13,18 +13,18 @@ import java.util.List;
 
 @Getter
 @Setter
-public class TestCase {
+public class ApprovalTestCase {
 
     public String description;
     public String input;
     public String expected;
 
-    public static List<TestCase> deserializeTestCases(String path) throws IOException {
+    public static List<ApprovalTestCase> deserializeTestCases(String path) throws IOException {
         return deserializeTestCases(path, "");
     }
 
-    public static List<TestCase> deserializeTestCases(String path, String startFlag) throws IOException {
-        List<TestCase> cases = new ArrayList<>();
+    public static List<ApprovalTestCase> deserializeTestCases(String path, String startFlag) throws IOException {
+        List<ApprovalTestCase> cases = new ArrayList<>();
         File file = new File(path);
         byte[] jsonData = Files.readAllBytes(file.toPath());
         JSONObject jsonObject = new JSONObject(new String(jsonData));
@@ -32,17 +32,17 @@ public class TestCase {
         JSONArray jsonArray = jsonObject.getJSONArray("cases");
 
         for (int i = 0; i < jsonArray.length(); i++) {
-            TestCase testCase = new TestCase();
+            ApprovalTestCase approvalTestCase = new ApprovalTestCase();
             JSONObject json = jsonArray.getJSONObject(i);
 
-            testCase.setDescription(json.getString("description"));
+            approvalTestCase.setDescription(json.getString("description"));
 
             JSONObject input = json.getJSONObject("input");
-            testCase.setInput(startFlag + input.toString()); // Add the 2-byte length prefix to the input
+            approvalTestCase.setInput(startFlag + input.toString()); // Add the 2-byte length prefix to the input
 
-            testCase.setExpected(json.get("expected").toString());
+            approvalTestCase.setExpected(json.get("expected").toString());
 
-            cases.add(testCase);
+            cases.add(approvalTestCase);
         }
         return cases;
     }
